version 1 features and functionality complete
Tube version 1 features and functionality are now complete. After 8 movies have been uploaded to the filesystem, the app will begin deleting the oldest movie from the filesystem and database as each new movie is uploaded. In essence, the app will now maintain only the 8 most recently uploaded videos. This feature was added for maintainability purposes, as server space is expensive!